Created qaoa_cicuit_utils file.
